% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfa_fit_plot.R
\name{cfa_fit_plot}
\alias{cfa_fit_plot}
\title{Fit and plot CFA simultaneously}
\usage{
cfa_fit_plot(model, data, covs = FALSE, remove.items = "", file.name)
}
\arguments{
\item{model}{CFA model to fit.}

\item{data}{Data set on which to fit the CFA model.}

\item{covs}{Logical, whether to include covariances on the lavaan plot.}

\item{remove.items}{Optional, if one wants to remove items from the CFA model
without having to redefine it completely again.}

\item{file.name}{Optional, if one wants something different than the default
file name. Defaults to saving to the "/model" subfolder of
the working directory. If it doesn't exist, it creates it.}
}
\description{
Print and save CAF fit, as well as plot CFA loadings, simultaneously.
}
\examples{
(latent <- list(visual = paste0("x", 1:3),
                textual = paste0("x", 4:6),
                speed = paste0("x", 7:9)))

HS.model <- write_lavaan(latent = latent)
cat(HS.model)

library(lavaan)
fit <- cfa_fit_plot(HS.model, HolzingerSwineford1939)
}
\keyword{CFA,}
\keyword{fit}
\keyword{lavaan,}
\keyword{plot,}
