% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discover_indirect_effects.R
\name{.dfs_find_paths}
\alias{.dfs_find_paths}
\title{Depth-first search to find paths}
\usage{
.dfs_find_paths(current_node, edges, max_depth, current_path, visited)
}
\arguments{
\item{current_node}{current node in traversal}

\item{edges}{edge dataframe}

\item{max_depth}{maximum search depth}

\item{current_path}{current path being built}

\item{visited}{nodes already visited in current path}
}
\value{
list of paths
}
\description{
Depth-first search to find paths
}
