% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discover_indirect_effects.R
\name{.find_all_indirect_paths}
\alias{.find_all_indirect_paths}
\title{Find all indirect paths in the graph}
\usage{
.find_all_indirect_paths(
  graph_info,
  max_chain_length = 5,
  min_chain_length = 2,
  computational_limit = 1000,
  verbose = FALSE
)
}
\arguments{
\item{graph_info}{list with nodes and edges}

\item{max_chain_length}{maximum path length}

\item{min_chain_length}{minimum path length}

\item{computational_limit}{maximum paths to find}

\item{verbose}{whether to print progress}
}
\value{
list of indirect paths
}
\description{
Find all indirect paths in the graph
}
