% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discover_indirect_effects.R
\name{.generate_indirect_syntax}
\alias{.generate_indirect_syntax}
\title{Generate lavaan syntax for indirect effects}
\usage{
.generate_indirect_syntax(indirect_paths)
}
\arguments{
\item{indirect_paths}{list of paths}
}
\value{
named list of indirect effect definitions
}
\description{
Generate lavaan syntax for indirect effects
}
