% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discover_indirect_effects.R
\name{.parse_covariance_line}
\alias{.parse_covariance_line}
\title{Parse covariance line into edges}
\usage{
.parse_covariance_line(line)
}
\arguments{
\item{line}{covariance line from lavaan syntax}
}
\value{
list with edges and nodes
}
\description{
Parse covariance line into edges
}
