% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discover_indirect_effects.R
\name{.parse_latent_line}
\alias{.parse_latent_line}
\title{Parse latent variable line into edges}
\usage{
.parse_latent_line(line)
}
\arguments{
\item{line}{latent variable line from lavaan syntax}
}
\value{
list with edges and nodes
}
\description{
Parse latent variable line into edges
}
