% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discover_indirect_effects.R
\name{.parse_regression_line}
\alias{.parse_regression_line}
\title{Parse regression line into edges}
\usage{
.parse_regression_line(line)
}
\arguments{
\item{line}{regression line from lavaan syntax}
}
\value{
list with edges and nodes
}
\description{
Parse regression line into edges
}
