% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discover_indirect_effects.R
\name{.parse_variable_list}
\alias{.parse_variable_list}
\title{Parse variable list with potential labels}
\usage{
.parse_variable_list(var_string)
}
\arguments{
\item{var_string}{string containing variables and labels}
}
\value{
list with variables and labels
}
\description{
Parse variable list with potential labels
}
