% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaanExtra-package.R
\docType{package}
\name{lavaanExtra-package}
\alias{lavaanExtra}
\alias{lavaanExtra-package}
\title{lavaanExtra: Convenience Functions for Package 'lavaan'}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Affords an alternative, vector-based syntax to 'lavaan', as well as other convenience functions such as naming paths and defining indirect links automatically, in addition to convenience formatting optimized for a publication and script sharing workflow.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://lavaanExtra.remi-theriault.com}
  \item Report bugs at \url{https://github.com/rempsyc/lavaanExtra/issues}
}

}
\author{
\strong{Maintainer}: Rémi Thériault \email{remi.theriault@mail.mcgill.ca} (\href{https://orcid.org/0000-0003-4315-6788}{ORCID})

}
\keyword{internal}
