% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopyOfcfa_fit_plot.R
\name{foo}
\alias{foo}
\title{Test function to print lavaan fit summary}
\usage{
foo(fit)
}
\arguments{
\item{fit}{model to fit.}
}
\description{
NA
}
\examples{
HS.model <- 'visual =~ x1 + x2 + x3'
library(lavaan)
fit <- lavaan::cfa(model = HS.model, data = HolzingerSwineford1939)
foo(fit)
}
\keyword{CFA,}
\keyword{fit}
\keyword{lavaan,}
\keyword{plot,}
