% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaan_cov.R
\name{lavaan_cov}
\alias{lavaan_cov}
\title{Extract relevant indirect effects indices from lavaan model}
\usage{
lavaan_cov(fit, nice_table = FALSE, ...)
}
\arguments{
\item{fit}{lavaan fit object to extract fit indices from}

\item{nice_table}{Logical, whether to print the table as a \code{rempsyc::nice_table}
as well as print the reference values at the bottom of the table.}

\item{...}{Arguments to be passed to \code{rempsyc::nice_table}}
}
\description{
Extract relevant indirect effects indices from lavaan model.
}
\examples{

(latent <- list(visual = paste0("x", 1:3),
                textual = paste0("x", 4:6),
                speed = paste0("x", 7:9)))

(regression <- list(ageyr = c("visual", "textual", "speed"),
                    grade = c("visual", "textual", "speed")))

(mediation <- list(speed = "visual",
                   textual = "visual",
                   visual = c("ageyr", "grade")))

(indirect <- list(age_visual_speed = c("speed_a", "visual_a"),
                  grade_visual_textual = c("textual_a", "visual_b")))

HS.model <- write_lavaan(mediation, regression, indirect = indirect,
                         latent = latent, label = TRUE)
cat(HS.model)

library(lavaan)
fit <- lavaan(HS.model, data=HolzingerSwineford1939,
              auto.var=TRUE, auto.fix.first=TRUE,
              auto.cov.lv.x=TRUE)
lavaan_cov(fit)
}
\keyword{CFA}
\keyword{analysis,}
\keyword{equation}
\keyword{lavaan,}
\keyword{modeling,}
\keyword{path}
\keyword{structural}
