% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaan_ind.R
\name{lavaan_ind}
\alias{lavaan_ind}
\title{Extract relevant indirect effects indices from lavaan model}
\usage{
lavaan_ind(fit, nice_table = FALSE, ...)
}
\arguments{
\item{fit}{lavaan fit object to extract fit indices from}

\item{nice_table}{Logical, whether to print the table as a
\code{rempsyc::nice_table} as well as print the
reference values at the bottom of the table.}

\item{...}{Arguments to be passed to \code{rempsyc::nice_table}}
}
\value{
A dataframe, including the indirect effect, corresponding paths,
standardized regression coefficient, and corresponding p-value.
}
\description{
Extract relevant indirect effects indices from lavaan model
through \verb{lavaan::parameterEstimates)} with
\code{standardized = TRUE}. In this case, the beta (B) represents
the resulting \code{std.all} column.
}
\examples{

(latent <- list(visual = paste0("x", 1:3),
                textual = paste0("x", 4:6),
                speed = paste0("x", 7:9)))

(mediation <- list(speed = "visual",
                   textual = "visual",
                   visual = c("ageyr", "grade")))

(indirect <- list(IV = c("ageyr", "grade"),
                  M = "visual",
                  DV = c("speed", "textual")))

HS.model <- write_lavaan(mediation, indirect = indirect,
                         latent = latent, label = TRUE)
cat(HS.model)

library(lavaan)
fit <- lavaan(HS.model, data=HolzingerSwineford1939,
              auto.var=TRUE, auto.fix.first=TRUE,
              auto.cov.lv.x=TRUE)
lavaan_ind(fit)
}
\keyword{CFA}
\keyword{analysis}
\keyword{equation}
\keyword{lavaan}
\keyword{modeling}
\keyword{path}
\keyword{structural}
