% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaan_reg.R
\name{lavaan_reg}
\alias{lavaan_reg}
\title{Vector-based lavaan syntax interpreter}
\usage{
lavaan_reg(fit)
}
\arguments{
\item{fit}{lavaan fit object to extract fit indices from}
}
\description{
Vector-based lavaan syntax interpreter.
}
\examples{
(latent <- list(visual = paste0("x", 1:3),
                textual = paste0("x", 4:6),
                speed = paste0("x", 7:9)))

(regression <- list(ageyr = c("visual", "textual", "speed"),
                    grade = c("visual", "textual", "speed")))

HS.model <- write_lavaan(latent = latent, regression = regression)
cat(HS.model)

library(lavaan)
fit <- lavaan(HS.model, data=HolzingerSwineford1939,
              auto.var=TRUE, auto.fix.first=TRUE,
              auto.cov.lv.x=TRUE)
lavaan_reg(fit)
}
\keyword{CFA}
\keyword{analysis,}
\keyword{equation}
\keyword{lavaan,}
\keyword{modeling,}
\keyword{path}
\keyword{structural}
