% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_as_x.R
\name{save_as_x}
\alias{save_as_x}
\alias{save_as_docx}
\alias{save_as_html}
\alias{save_as_image}
\alias{save_as_pptx}
\title{\code{flextable::save_as_x}}
\usage{
save_as_docx(...)

save_as_html(...)

save_as_image(...)

save_as_pptx(...)
}
\arguments{
\item{...}{arguments passed to \code{save_as_x} functions}
}
\value{
A table, either as an image, Word, PowerPoint, or HTML document.
}
\description{
These \code{save_as_x} functions are imported from \code{flextable}.
}
