% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_lavaan.R
\name{write_lavaan}
\alias{write_lavaan}
\title{Vector-based lavaan syntax interpreter}
\usage{
write_lavaan(
  mediation = NULL,
  regression = NULL,
  covariance = NULL,
  indirect = NULL,
  latent = NULL,
  intercept = NULL,
  label = FALSE,
  use.letters = FALSE
)
}
\arguments{
\item{mediation}{Mediation    indicators (\code{~} symbol: "is regressed on"). Differs from
argument \code{regression} because path names can be optionally
specified automatically with argument \code{label}.}

\item{regression}{Regression indicators (\code{~} symbol: "is regressed on").}

\item{covariance}{(Residual) (co)variance indicators (\verb{~~} symbol: "is correlated with").}

\item{indirect}{Indirect effect indicators (\verb{:=} symbol: "indirect effect defined as").}

\item{latent}{Latent variable indicators (\verb{=~} symbol: "is measured by").}

\item{intercept}{Intercept indicators (\code{~ 1} symbol: "intercept").}

\item{label}{Logical, whether to display path names for the mediation argument.}

\item{use.letters}{Logical, for the labels, whether to use letters instead of the variable names.}
}
\description{
Vector-based lavaan syntax interpreter.
}
\examples{
(latent <- list(visual = paste0("x", 1:3),
                textual = paste0("x", 4:6),
                speed = paste0("x", 7:9)))

HS.model <- write_lavaan(latent = latent)
cat(HS.model)

library(lavaan)
fit <- lavaan(HS.model, data=HolzingerSwineford1939,
              auto.var=TRUE, auto.fix.first=TRUE,
              auto.cov.lv.x=TRUE)
summary(fit, fit.measures=TRUE)
}
\keyword{CFA}
\keyword{analysis,}
\keyword{equation}
\keyword{lavaan,}
\keyword{modeling,}
\keyword{path}
\keyword{structural}
